Organized optional params, created range type
The Gauge constructor function now takes an id representing
the div to render the gauge to, along with an options object
literal. The following optional parameters are supported in
options.

name
    The name of the gauge.

min
    The minimum range extent of the dial.

max
    The maximum range extent of the dial.

nominal

    The point on the dial representing the nominal value.

nominalMin
    The minimum range of the nominal dial extent

nominal Max
    The maximum range of the nominal dial extent

numMinorDivisions
    The amount of minor divisions between major ticks

angleSpan
    The degrees in which the gauge face should subtend.

gaugeCenter
    The size of the gauge.

In addition the type range was created to handle the case of a min/max
constraint.